Support to enable/disable the Bugzilla integration per project
It can now be configured per project whether the Bugzilla integration is
enabled or not. To enable the Bugzilla integration for a project the
project must have the following entry in its `project.config` file in
the `refs/meta/config` branch:

```
  [plugin "bugzilla"]
    enabled = true
```

If `plugin.bugzilla.enabled` is not specified in the `project.config` file
the value is inherited from the parent project. If it is also not set
on any parent project the Bugzilla integration is disabled for this
project.

By setting `plugin.bugzilla.enabled` to true in the `project.config` of the
`All-Projects` project the Bugzilla integration can be enabled by default
for all projects. During the initialization of the plugin you are asked
if the Bugzilla integration should be enabled by default for all projects
and if yes this setting in the `project.config` of the `All-Projects`
project is done automatically.

This uses new functionality in Gerrit and hence the Gerrit API
version needs to be changed to 2.9-SNAPSHOT.

Change-Id: I7afe1c6fdbd60df283415ef36cd819ecfdc0c3b0
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>